Separate LambdaWrapper to more functions

+++ test/integration/utils/mongodb.integration.utils.js
@@ -5,15 +5,17 @@ module.exports.insert = (doc, mongodb) => {
         const dbName = 'testDB';
         const collectionName = 'testCollection';
         MongoClient.connect(mongoURL, { useNewUrlParser: true }, (err, client) => {
-            const db = client.db(dbName);
-            const collection = db.collection(collectionName);
-            collection.insertOne(doc)
-                .then((res) => client.close())
-                .then(() => resolve())
-                .catch((err) => {
-                    client.close()
-                        .then(resolve);
-                });
+            if (client) {
+                const db = client.db(dbName);
+                const collection = db.collection(collectionName);
+                collection.insertOne(doc)
+                    .then((res) => client.close())
+                    .then(() => resolve())
+                    .catch((err) => {
+                        client.close()
+                            .then(resolve);
+                    });
+            }
         });
     });
 };
@@ -25,17 +27,19 @@ module.exports.update = (doc, updateKey, mongodb) => {
         const dbName = 'testDB';
         const collectionName = 'testCollection';
         MongoClient.connect(mongoURL, { useNewUrlParser: true }, (err, client) => {
-            const db = client.db(dbName);
-            const collection = db.collection(collectionName);
-            collection.insertOne(doc)
-                .then((res) => collection.updateOne(updateKey, {
-                    '$set': { 'newField': 'newValue ' },
-                })).then((res) => client.close())
-                .then(() => resolve())
-                .catch((err) => {
-                    client.close()
-                        .then(resolve);
-                });
+            if (client) {
+                const db = client.db(dbName);
+                const collection = db.collection(collectionName);
+                collection.insertOne(doc)
+                    .then((res) => collection.updateOne(updateKey, {
+                        '$set': { 'newField': 'newValue ' },
+                    })).then((res) => client.close())
+                    .then(() => resolve())
+                    .catch((err) => {
+                        client.close()
+                            .then(resolve);
+                    });
+            }
         });
     });
 };
@@ -47,14 +51,16 @@ module.exports.dropCollection = (mongodb) => {
         const dbName = 'testDB';
         const collectionName = 'testCollection';
         MongoClient.connect(mongoURL, { useNewUrlParser: true }, (err, client) => {
-            const db = client.db(dbName);
-            db.dropCollection('non_exist')
-                .then((res) => client.close())
-                .then(() => resolve())
-                .catch((err) => {
-                    client.close()
-                        .then(resolve);
-                });
+            if (client) {
+                const db = client.db(dbName);
+                db.dropCollection('non_exist')
+                    .then((res) => client.close())
+                    .then(() => resolve())
+                    .catch((err) => {
+                        client.close()
+                            .then(resolve);
+                    });
+            }
         });
     });
 };
